+++ daq/writer/include/emu/daq/writer/RateLimiter.h
@@ -1,4 +1,5 @@
 #include <sys/time.h>
+#include <stdint.h>
 
 namespace emu{
   namespace daq{
@@ -10,14 +11,14 @@ namespace emu{
       /// for the rate to be below \ref limitInHz, it rejects all subsequent events until that time 
       /// (that is, until \ref endOfVetoPeriod_). Then it starts accepting them again.
       class RateLimiter{
-	int             limitInHz_;             ///< max rate in Hz
-	int             sampleSize_;            ///< number of events in the sample
-	int             minTimeSpanInMicrosec_; ///< min time the sample must span for the rate to be below the limit
+	uint64_t        limitInHz_;             ///< max rate in Hz
+	uint64_t        sampleSize_;            ///< number of events in the sample
+	uint64_t        minTimeSpanInMicrosec_; ///< min time the sample must span for the rate to be below the limit
 	struct timeval *timeStamps_;            ///< a ring buffer of time stamps and of size sampleSize_
-	int             firstTimeIndex_;        ///< index of the oldest time stamp in the timeStamps_ array
-	int             lastTimeIndex_;         ///< index of the most recent time stamp in the timeStamps_ array
+	uint64_t        firstTimeIndex_;        ///< index of the oldest time stamp in the timeStamps_ array
+	uint64_t        lastTimeIndex_;         ///< index of the most recent time stamp in the timeStamps_ array
 	struct timeval  endOfVetoPeriod_;       ///< all events are rejected until the time specified by this
-	int             eventCount_;            ///< number of events since start or last lift of veto
+	uint64_t        eventCount_;            ///< number of events since start or last lift of veto
 	bool            debug_;
 
 	struct timeval  now_;                   ///< gettimeofday will set this to the current time
@@ -29,7 +30,7 @@ namespace emu{
 	/// @param limitInHz rate limit in Hertz; set it to 0 for no limit
 	/// @param sampleSize sample size (number of events)
 	///
-	RateLimiter( const int limitInHz, const int sampleSize );
+	RateLimiter( const uint64_t limitInHz, const uint64_t sampleSize );
 
 	/// Destructor.
 	///
